pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bithumbtradekit"
version = "0.1.0"
description = "Bithumb API 2.0 기반 코인 매매 도구"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "infinist", email = "infinist7@naver.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["bithumb", "cryptocurrency", "trading", "api"]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.25.0",
    "pandas>=1.3.0",
    "PyJWT>=2.4.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.900",
]

[project.urls]
Homepage = "https://github.com/infinist7/bithumbtradekit"
Repository = "https://github.com/infinist7/bithumbtradekit"
Issues = "https://github.com/infinist7/bithumbtradekit/issues"

[project.scripts]
bithumbtradekit = "bithumbtradekit.cli:main"
bt = "bithumbtradekit.cli:main"

[tool.hatch.version]
path = "src/bithumbtradekit/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["src/bithumbtradekit"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]